+++ src/renderer/containers/App.tsx
@@ -9,11 +9,9 @@ import Connect from '@renderer/containers/Connect';
 import Layout from '@renderer/containers/Layout';
 import {connect} from '@renderer/dispatchers/app';
 import {getActiveBank, getActiveBankConfig, getManagedAccounts} from '@renderer/selectors';
-import {AppDispatch, NotificationType} from '@renderer/types';
+import {AppDispatch} from '@renderer/types';
 import {formatSocketAddress} from '@renderer/utils/address';
-import initializeSockets from '@renderer/utils/sockets';
-import handleConfirmationBlockNotification from '@renderer/utils/sockets/confirmation-block-notifications';
-import {displayErrorToast} from '@renderer/utils/toast';
+import {initializeSockets, processSocketEvent} from '@renderer/utils/sockets';
 
 const App: FC = () => {
   const [loading, setLoading] = useState<boolean>(true);
@@ -35,14 +33,7 @@ const App: FC = () => {
 
     sockets.forEach((socket: any) => {
       socket.onmessage = (event: any) => {
-        try {
-          const notification = JSON.parse(event.data);
-          if (notification.notification_type === NotificationType.confirmationBlockNotification) {
-            handleConfirmationBlockNotification(accountNumbers, dispatch, notification);
-          }
-        } catch (error) {
-          displayErrorToast(error);
-        }
+        processSocketEvent(accountNumbers, dispatch, event);
       };
     });
